pyproject.toml
[tool.poetry]
name = "rutracker-trends"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
#packages = [{include = "rutracker_trends"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.28.2"
beautifulsoup4 = "^4.11.2"
pandas = "^1.5.3"
notebook = "^6.5.3"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
